+++ packages/awino-ui/src/lib/helpers.ts
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { ID } from '@/types/app';
 
 export function tabA11yProps(id: ID, index: number) {
@@ -6,3 +8,13 @@ export function tabA11yProps(id: ID, index: number) {
     'aria-controls': `tabpanel-${id}-${index}`,
   };
 }
+
+/**
+ * Wrap callback method, with event stop propagation
+ */
+export const stopPropagation =
+  (callback: () => void) =>
+  (event: React.SyntheticEvent): void => {
+    event.stopPropagation();
+    callback();
+  };